Move base measurements to DeviceMeasurement and update to latest telematics contracts
