+++ MojioSDK/Models/Battery.swift
@@ -9,12 +9,8 @@
 import UIKit
 import ObjectMapper
 
-public class BatteryVoltageUnits : NSObject {
-    public static let MilliVolts : String = "MilliVolts"
-    public static let Volts : String = "Volts"
-}
-
-public class Battery : Mappable {
+// Units in BatteryVoltageUnits
+public class Battery : DeviceMeasurement {
     
     public dynamic var Connected : Bool = false
     
@@ -22,32 +18,17 @@ public class Battery : Mappable {
     public dynamic var RiskSeverity : String? = nil
     public var LowVoltageDuration : TimePeriod? = nil
     public var HighVoltageDuration : TimePeriod? = nil
-    public dynamic var BaseUnit : String? = nil
-    public dynamic var Timestamp : String? = nil
-    public dynamic var BaseValue : Float = 0
-    
-    // BatteryVoltageUnits
-    public dynamic var Unit : String? = nil
-    public dynamic var Value : Float = 0
     
     public required convenience init?(_ map: Map) {
         self.init()
     }
     
-    public required init() {
-        
-    }
-    
-    public func mapping(map: Map) {
+    public override func mapping(map: Map) {
+        super.mapping(map)
+
         Connected <- map["Connected"]
         RiskSeverity <- map["RiskSeverity"]
         LowVoltageDuration <- map["LowVoltageDuration"]
         HighVoltageDuration <- map["HighVoltageDuration"]
-        BaseUnit <- map["BaseUnit"]
-        Timestamp <- map["Timestamp"]
-        BaseValue <- map["BaseValue"]
-        Unit <- map["Unit"]
-        Value <- map["Value"]
     }
-
 }
