+++ MojioSDK/Models/Distance.swift
@@ -9,24 +9,11 @@
 import UIKit
 import ObjectMapper
 
-public class Distance: Mappable {
-    
-    public dynamic var BaseUnit : String? = nil
-    public dynamic var SpeedBandId : Int = 0
-    public var SpeedBandDuration : TimePeriod? = nil
-    public dynamic var Timestamp  : String? = nil
-    public dynamic var BaseValue : Float = 0
-    
-    // SpeedUnits
-    public dynamic var Unit  : String? = nil
-    public dynamic var Value : Float = 0
+// Units in DistanceUnits
+public class Distance: DeviceMeasurement {
     
     public required convenience init?(_ map: Map) {
         self.init()
-    }
-    
-    public required init() {
-        
     }
     
     public func jsonDict() -> NSDictionary {
@@ -44,15 +31,4 @@ public class Distance: Mappable {
 
         return dictionary
     }
-    
-    public func mapping(map: Map) {
-        BaseUnit <- map["BaseUnit"]
-        SpeedBandId <- map["SpeedBandId"]
-        SpeedBandDuration <- map["SpeedBandDuration"]
-        Timestamp <- map["Timestamp"]
-        BaseValue <- map["BaseValue"]
-        Unit <- map["Unit"]
-        Value <- map["Value"]
-    }
-    
 }
