+++ MojioSDK/Models/FuelEfficiency.swift
@@ -9,42 +9,10 @@
 import UIKit
 import ObjectMapper
 
-public class FuelEfficiencyUnits : NSObject {
-    public static let MilesPerGallon : String = "MilesPerGallon"
-    public static let KilometerPerLiter : String = "KilometerPerLiter"
-    public static let LitersPerHundredKilometers : String = "LitersPerHundredKilometers"
-}
-
-public class FuelEfficiency: Mappable {
-    
-    public dynamic var BaseUnit : String? = nil
-    public var Statistics : MeasurementStatistics? = nil
-    public dynamic var BenchmarkTime : String? = nil
-    public var BenchmarkStatistics : MeasurementStatistics? = nil
-    public dynamic var Timestamp : String? = nil
-    public dynamic var BaseValue : Float = 0
-    
-    // FuelEfficiencyUnits
-    public dynamic var Unit  : String? = nil
-    public dynamic var Value : Float = 0
+// Units in FuelEfficiencyUnits
+public class FuelEfficiency: DeviceMeasurement {
 
     public required convenience init?(_ map: Map) {
         self.init()
     }
-    
-    public required init() {
-        
-    }
-
-    public func mapping(map: Map) {
-        BaseUnit <- map["BaseUnit"]
-        Statistics <- map["Statistics"]
-        BenchmarkTime <- map["BenchmarkTime"]
-        BenchmarkStatistics <- map["BenchmarkStatistics"]
-        Timestamp <- map["Timestamp"]
-        BaseValue <- map["BaseValue"]
-        Unit <- map["Unit"]
-        Value <- map["Value"]
-    }
-
 }
