+++ MojioSDK/Models/FuelLevel.swift
@@ -9,36 +9,20 @@
 import UIKit
 import ObjectMapper
 
-public class FuelLevelUnits : NSObject {
-    public static let Percentage : String = "Percentage"
-}
 
-public class FuelLevel: Mappable {
+// Units are in PercentageUnits
+public class FuelLevel: DeviceMeasurement {
     
-    public dynamic var BaseUnit : String? = nil
     public dynamic var RiskSeverity : String? = nil
-    public dynamic var Timestamp : String? = nil
-    public dynamic var BaseValue : Float = 0
-    
-    // FuelLevelUnits
-    public dynamic var Unit : String? = nil
-    public dynamic var Value : Float = 0
     
     public required convenience init?(_ map: Map) {
         self.init()
     }
     
-    public required init() {
-        
-    }
+    public override func mapping(map: Map) {
+        super.mapping(map)
 
-    public func mapping(map: Map) {
-        BaseUnit <- map["BaseUnit"]
         RiskSeverity <- map["RiskSeverity"]
-        Timestamp <- map["Timestamp"]
-        BaseValue <- map["BaseValue"]
-        Unit <- map["Unit"]
-        Value <- map["Value"]
     }
 
 }
