+++ MojioSDK/Models/FuelVolume.swift
@@ -9,32 +9,11 @@
 import UIKit
 import ObjectMapper
 
-public class FuelVolumeUnits : NSObject {
-    public static let Gallon        : String = "Gallon"
-    public static let Liter         : String = "Liter"
-    public static let Gill          : String = "Gill"
-    public static let Pint          : String = "Pint"
-    public static let Quart         : String = "Quart"
-    public static let Milliliter    : String = "Milliliter"
-}
 
-public class FuelVolume: Mappable {
+// Units in FuelVolumeUnits
+public class FuelVolume: DeviceMeasurement {
 
-    public dynamic var BaseUnit : String? = nil
-    public dynamic var Timestamp : String? = nil
-    public dynamic var BaseValue : Float = 0
-    public dynamic var Unit : String? = nil
-    public dynamic var Value : Float = 0
-    
     public required convenience init?(_ map: Map) {
         self.init()
-    }
-    
-    public func mapping(map: Map) {
-        BaseUnit <- map["BaseUnit"];
-        Timestamp <- map["Timestamp"];
-        BaseValue <- map["BaseValue"];
-        Unit <- map["Unit"];
-        Value <- map["Value"];
     }
 }
