+++ MojioSDK/Models/RPM.swift
@@ -9,34 +9,10 @@
 import UIKit
 import ObjectMapper
 
-public class RPMUnits {
-    public static let RevolutionsPerMinute : String = "RevolutionsPerMinute"
-}
-
-public class RPM: Mappable {
-    
-    public dynamic var BaseUnit : String? = nil
-    public dynamic var Timestamp : String? = nil
-    public dynamic var BaseValue : Float = 0
-    
-    // RPMUnits
-    public dynamic var Unit : String? = nil
-    public dynamic var Value : Float = 0
+// Units in RPMUnits
+public class RPM: DeviceMeasurement {
     
     public required convenience init?(_ map: Map) {
         self.init()
     }
-    
-    public required init() {
-        
-    }
-
-    public func mapping(map: Map) {
-        BaseUnit <- map["BaseUnit"]
-        Timestamp <- map["Timestamp"]
-        BaseValue <- map["BaseValue"]
-        Unit <- map["Unit"]
-        Value <- map["Value"]
-    }
-
 }
