+++ MojioSDK/Models/Speed.swift
@@ -9,31 +9,13 @@
 import UIKit
 import ObjectMapper
 
-public class SpeedUnits {
-    public static let KilometersPerHour : String = "KilometersPerHour"
-    public static let MilesPerHour : String = "MilesPerHour"
-}
-
-public class Speed: Mappable {
-    
-    public dynamic var BaseUnit : String? = nil
-    public dynamic var SpeedBandId : Int = 0
-    public var SpeedBandDuration : TimePeriod? = nil
-    public dynamic var Timestamp  : String? = nil
-    public dynamic var BaseValue : Float = 0
-    
-    // SpeedUnits
-    public dynamic var Unit  : String? = nil
-    public dynamic var Value : Float = 0
+// Units in SpeedUnits
+public class Speed: DeviceMeasurement {
     
     public required convenience init?(_ map: Map) {
         self.init()
     }
     
-    public required init() {
-        
-    }
-
     public func jsonDict () -> NSDictionary {
         let dictionary : NSMutableDictionary = NSMutableDictionary()
         
@@ -45,15 +27,4 @@ public class Speed: Mappable {
         
         return dictionary
     }
-    
-    public func mapping(map: Map) {
-        BaseUnit <- map["BaseUnit"]
-        SpeedBandId <- map["SpeedBandId"]
-        SpeedBandDuration <- map["SpeedBandDuration"]
-        Timestamp <- map["Timestamp"]
-        BaseValue <- map["BaseValue"]
-        Unit <- map["Unit"]
-        Value <- map["Value"]
-    }
-
 }
