+++ MojioSDK/Models/TimePeriod.swift
@@ -9,42 +9,10 @@
 import UIKit
 import ObjectMapper
 
-public class TimePeriodUnits : NSObject {
-    public static let Ticks : String = "Ticks"
-    public static let Milliseconds : String = "Milliseconds"
-    public static let Seconds : String = "Seconds"
-    public static let Minutes : String = "Minutes"
-    public static let Hours : String = "Hours"
-    public static let Days : String = "Days"
-    public static let Weeks : String = "Weeks"
-    public static let Months : String = "Months"
-    public static let Years : String = "Years"
-}
-
-public class TimePeriod: Mappable {
-    
-    public dynamic var BaseUnit : String? = nil
-    public dynamic var Timestamp : String? = nil
-    public dynamic var BaseValue : Float = 0
-    
-    // TimePeriodUnits
-    public dynamic var Unit : String? = nil
-    public dynamic var Value : Float = 0
+// Units in TimePeriodUnits
+public class TimePeriod: DeviceMeasurement {
     
     public required convenience init?(_ map: Map) {
         self.init()
     }
-    
-    public required init() {
-        
-    }
-
-    public func mapping(map: Map) {
-        BaseUnit <- map["BaseUnit"]
-        Timestamp <- map["Timestamp"]
-        BaseValue <- map["BaseValue"]
-        Unit <- map["Unit"]
-        Value <- map["Value"]
-    }
-
 }
